[Sprint 48] S48-1: rename `aimx status` to `aimx doctor`
Clean rename — the diagnostic command lines up with the `brew doctor` /
`flutter doctor` / `npm doctor` shape. Pre-launch so no alias is kept;
`aimx status` now produces the standard clap "unrecognized subcommand"
error.

- src/status.rs renamed to src/doctor.rs (mod + dispatch wired through)
- Command::Status removed from cli.rs; Command::Doctor takes its place
- Repo-wide sweep: README.md, CLAUDE.md, book/{configuration,
  getting-started,setup,troubleshooting}.md, docs/manual-{test,setup}.md
- Existing status integration tests renamed; new test asserts that
  `aimx status` fails with a clap error so the rename can't silently
  regress
