+++ book/getting-started.md
@@ -84,8 +84,8 @@ After DNS records propagate, verify the setup:
 # Check port 25 connectivity (requires root)
 sudo aimx portcheck
 
-# Check server status, mailbox counts, and DNS verification
-aimx status
+# Check server health, mailbox counts, and DNS verification
+aimx doctor
 ```
 
 ## Send a test email
